refactor: Introduce FilterButtonStyleHelper and move FilterImageName
